requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.6.0
Brotli==1.0.9
certifi==2019.6.16
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.41.0
django-ckeditor==6.0.0
django-cogwheels==0.2
django-compression-middleware==0.4.1
django-enumchoicefield==2.0.0
django-filter==2.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-modelcluster==5.0.2
django-recaptcha==2.0.6
django-taggit==1.3.0
django-treebeard==4.3
djangorestframework==3.11.1
draftjs-exporter==2.1.6
et-xmlfile==1.0.1
fanstatic==1.1
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==3.4.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
js.jquery==3.3.1
js.knockout==3.1.0
l18n==2018.5
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.7.7
psycopg2-binary==2.8.3
psycopg2-pool==1.1
pycparser==2.19
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.12.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.25.3
virtualenv==20.3.0
wagtail==2.8.2
wagtail-condensedinlinepanel==0.5.2
wagtail-django-recaptcha==1.0
wagtail-robots==0.1.2
wagtailmenus==2.12
webencodings==0.5.1
WebOb==1.8.5
whitenoise==5.2.0
Willow==1.3
zipp==3.4.0
zstandard==0.15.1
